services/api-gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
redis==5.0.4
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
sentry-sdk[fastapi]==2.1.1
structlog==24.1.0
python-dotenv==1.0.1
